refactor: sidebar 코드 분리
